Implemented mixed basis and fixed the H2P PERL debacle by implementing a completely Pythonic RegEx coordinate scraper.
